Fix: Tests can start up

+++ server/controllers/workspace.py
@@ -170,21 +170,20 @@ class AppFolderController:
             # Create new app folder
             create_folder(path=self.app_folder_path)
             create_init_file(path=self.app_folder_path)
-
             app_object = self._add_app_to_workspace_properties(
                 app_name=self.app_name, app_label=app_label
             )
-
             new_app_properties = self._get_app_properties()
             create_file(
                 path=self.app_folder_path,
                 content=json.dumps(new_app_properties, indent=2),
                 file_name="properties.json",
             )
+            response = None
             if router:
                 response = router.app.create_app(app_properties={**app_object})
 
-            if response.status_code != 200:
+            if response is not None and response.status_code != 200:
                 shutil.rmtree(self.app_folder_path)
                 self._remove_app_from_workspace_properties(self.app_name)
                 raise HTTPException(
@@ -409,7 +408,7 @@ class AppFolderController:
         existing_app_labels = [
             a["label"] for a in self._get_workspace_properties()["apps"]
         ]
-        if app_label in existing_app_labels:
+        if app_label is not None and app_label in existing_app_labels:
             raise HTTPException(
                 status_code=400, detail="Another app with the same label already exists"
             )
